Moved components into folder, styled button
