+++ src/App.js
@@ -1,8 +1,8 @@
 import './App.css';
 import './Setup.css';
-import Header from './Header';
-import GenreForm from './GenreForm';
-import Footer from './Footer'
+import Header from './components/Header';
+import GenreForm from './components/GenreForm';
+import Footer from './components/Footer'
